Caveman Repel ability
I made it so the player can now hold right click when near a person to push them away. I also hid the weaponry for now because that's not the focus. If you want to make it reappear, just delete the queue_free() function from the spear and club scripts.